--- src/main/java/roomescape/controller/api/GlobalExceptionRestHandler.java
@@ -1,30 +1,0 @@
-package roomescape.controller.api;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import java.time.format.DateTimeParseException;
-
-@RestControllerAdvice
-public class GlobalExceptionRestHandler {
-    // TODO 예외 처리 방식 통일 고려
-    @ExceptionHandler(RuntimeException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public String handleRuntimeException(RuntimeException e) {
-        return e.getMessage();
-    }
-
-    @ExceptionHandler(Exception.class)
-    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
-    public String handleCheckException(Exception e) {
-        return e.getMessage();
-    }
-
-    @ExceptionHandler(value = DateTimeParseException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public String handleException() {
-        return "올바르지 않은 날짜 / 시간 형식입니다.";
-    }
-}
