+++ src/main/java/roomescape/controller/web/admin/AdminPageController.java
@@ -1,0 +1,13 @@
+package roomescape.controller.web.admin;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class AdminPageController {
+
+    @GetMapping("/admin")
+    public String getAdminPage() {
+        return "/admin/index";
+    }
+}
