+++ src/main/resources/templates/error.html
@@ -63,8 +63,8 @@
 </nav>
 
 <div class="error-container">
-    <h1>404</h1>
-    <p>페이지를 찾을 수 없습니다.</p>
+    <h1>401</h1>
+    <p>페이지에 접근할 권한이 없습니다.</p>
 </div>
 
 <!--: [4단계] 사용자 로그인 - 로그인 기능 구현 후 활성화-->
